requirements.txt
python-dotenv
fastapi
sqlalchemy
asyncpg
uvicorn
greenlet
gunicorn